+++ integrations/slack/src/handlers/commands.ts
@@ -1,6 +1,6 @@
 import { FetchEventCallback } from '@gitbook/runtime';
 
-import { SlackRuntimeContext } from './configuration';
+import { SlackRuntimeContext } from '../configuration';
 
 export interface SlashEvent {
     /** Slack App's unique ID */
@@ -23,24 +23,31 @@ export interface SlashEvent {
      * Reference - https://api.slack.com/interactivity/handling#message_responses
      */
     response_url: string;
+
+    /** The name of the channel where the command was triggered */
+    channel_name: string;
+
+    /** The ID of the parent message, if any */
+    thread_ts?: string;
 }
 
 export function createSlackCommandsHandler(handlers: {
     [type: string]: (slashEvent: SlashEvent, context: SlackRuntimeContext) => Promise<any>;
 }): FetchEventCallback {
     return async (request, context) => {
-        const slashEvent = {} as SlashEvent;
-        new URLSearchParams(await request.text()).forEach((value, key) => {
-            slashEvent[key] = value;
-        });
+        const requestText = await request.text();
 
-        if (!slashEvent.command) {
+        const slashEvent: SlashEvent = Object.fromEntries(
+            new URLSearchParams(requestText).entries()
+        );
+
+        const { command } = slashEvent;
+
+        if (!command) {
             return new Response(`Invalid slash command`, {
                 status: 422,
             });
         }
-
-        const { command } = slashEvent;
 
         const handler = handlers[command];
         if (!handler) {
@@ -49,10 +56,8 @@ export function createSlackCommandsHandler(handlers: {
             });
         }
 
-        await handler(slashEvent, context);
+        const handlerPromise = handler(slashEvent, context);
 
-        return new Response(null, {
-            status: 200,
-        });
+        context.waitUntil(handlerPromise);
     };
 }