+++ integrations/slack/src/handlers/index.ts
@@ -1,0 +1,5 @@
+export * from './search';
+export * from './handlers';
+export * from './actions';
+export * from './commands';
+export * from './events';