+++ packages/cli/src/manifest.ts
@@ -27,6 +27,7 @@ export interface IntegrationManifest {
     externalLinks?: api.Integration['externalLinks'];
     organization?: string;
     secrets: { [key: string]: string };
+    entities?: api.RequestPublishIntegration['entities'];
     contentSecurityPolicy?: string;
 }
 
@@ -200,6 +201,12 @@ async function getManifestSchema() {
                 ...getAPIJsonSchemaFor(
                     openAPISpec,
                     'components/schemas/IntegrationContentSecurityPolicy'
+                ),
+            },
+            entities: {
+                ...getAPIJsonSchemaFor(
+                    openAPISpec,
+                    'components/schemas/RequestPublishIntegration/properties/entities'
                 ),
             },
             organization: {