+++ packages/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gitbook/api
 
+## 0.9.0
+
+### Minor Changes
+
+-   4affdac: Add ContentKit support for codeblock element for input
+
+## 0.8.0
+
+### Minor Changes
+
+-   691bc05: Bump API client to use the latest spec
+
 ## 0.7.0
 
 ### Minor Changes