+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gitbook/cli
 
+## 0.12.0
+
+### Minor Changes
+
+-   6109681: Update CLI to support entities while publishing an integration
+
+## 0.11.0
+
+### Minor Changes
+
+-   21891ed: Bump @gitbook/cli to use the latest API spec
+
 ## 0.10.0
 
 ### Minor Changes